api refactor: - '/crawls' create api accepts all options as POST body params, 'start' indicates if also starting - '/crawl/<id>/start' api simply starts crawl - 'cache' enum added to (always, default, never) to indicate aggressive cacheing (every request), default, or no cacheing - browsertrix_cli: move crawl commands to 'crawls' subcommand and crawls.py to allow other subcommands - tests: update tests for new api, fix name in test compose
