+++ .dockerignore
@@ -237,3 +237,4 @@ README.md
 mypy.ini
 .flake8
 frontend
+webarchive
