+++ Dockerfile
@@ -6,7 +6,8 @@ COPY requirements.txt ./
 
 RUN pip install --no-cache-dir -r requirements.txt
 
-COPY . ./
+COPY browsertrix ./browsertrix
+COPY static ./static
 
 CMD uvicorn --reload --host 0.0.0.0 --port 8000 browsertrix.api:app
