+++ browsertrix_cli/__init__.py
@@ -1,0 +1,6 @@
+import click
+
+# ============================================================================
+@click.group()
+def cli():
+    pass
